# perf: improve performance of new unresolved-reference rule